alx_travel_app/requirement.txt
# Core Django
Django==4.2.10  # LTS version with long-term support

# REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Async Task Processing
celery==5.3.6
pika==1.3.2  # RabbitMQ client for Python
django-celery-results==2.5.1  # Store Celery results in Django database
django-celery-beat==2.5.0  # Database-backed periodic tasks

# Database
mysqlclient==2.2.0

# Environment Configuration
django-environ==0.11.2